// template/react-starter-kit/src/.generated/api/model/noticeStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type NoticeStatus = typeof NoticeStatus[keyof typeof NoticeStatus];


export const NoticeStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  expired: 'expired',
} as const;